+++ packages/terra-clinical-data-grid/tests/wdio/data-grid-spec.js
@@ -155,13 +155,31 @@ Terra.describeViewports('DataGrid', ['medium', 'huge'], () => {
     Terra.validates.element('custom row/header heights', { selector: '#custom-height-data-grid' });
   });
 
-  describe('with highlighted column not selectable', () => {
-    before(() => {
-      browser.url('/raw/tests/terra-clinical-data-grid/clinical-data-grid/highlight-column-not-selectable-data-grid');
+  it('should display highlighted column not selectable cells DataGrid', () => {
+    browser.url('/raw/tests/terra-clinical-data-grid/clinical-data-grid/highlight-column-not-selectable-data-grid');
+
+    Terra.validates.element('highlighted column not selectable cells', { selector: '#highlight-column-not-selectable-data-grid' });
+  });
+
+  it('should display highlighted column with a single row DataGrid', () => {
+    browser.url('/raw/tests/terra-clinical-data-grid/clinical-data-grid/highlight-column-single-row-data-grid');
+
+    Terra.validates.element('highlighted column single row', { selector: '#highlight-column-single-row-data-grid' });
+  });
+
+  describe('with highlighted column empty sections', () => {
+    beforeEach(() => {
+      browser.url('/raw/tests/terra-clinical-data-grid/clinical-data-grid/highlight-column-empty-sections-data-grid');
     });
 
-    it('should display highlighted column not selectable cells DataGrid', () => {
-      Terra.validates.element('highlighted column not selectable cells', { selector: '#highlight-column-not-selectable-data-grid' });
+    it('should display highlighted column with a single row visible and collapsed and empty sections DataGrid', () => {
+      Terra.validates.element('highlighted column empty collapsed sections', { selector: '#highlight-column-empty-sections-data-grid' });
+    });
+
+    it('should display highlighted column with a decorative row DataGrid', () => {
+      $('[data-terra-clinical-data-grid-section-header-id="Section-2"] [class*="SectionHeader"]').click();
+
+      Terra.validates.element('highlighted column decorative row', { selector: '#highlight-column-empty-sections-data-grid' });
     });
   });
 
@@ -175,24 +193,24 @@ Terra.describeViewports('DataGrid', ['medium', 'huge'], () => {
     });
 
     it('should display highlighted column selectable cells DataGrid with row hover', () => {
-      $('[data-accessibility-id="23"]').moveTo();
+      $('[data-accessibility-id="24"]').moveTo();
       Terra.validates.element('highlighted column selectable cells row hover', { selector: '#highlight-column-selectable-data-grid' });
     });
 
     it('should display highlighted column selectable cells DataGrid with row selection', () => {
-      $('[data-accessibility-id="23"]').click();
+      $('[data-accessibility-id="24"]').click();
       $('#root').moveTo({ xoffset: 0, yoffset: 0 });
       Terra.validates.element('highlighted column selectable cells row selection', { selector: '#highlight-column-selectable-data-grid' });
     });
 
     it('should display highlighted column selectable cells DataGrid with cell focus', () => {
-      $('[data-accessibility-id="25"]').click();
+      $('[data-accessibility-id="26"]').click();
       browser.keys(['Tab']);
       Terra.validates.element('highlighted column selectable cells cell focus', { selector: '#highlight-column-selectable-data-grid' });
     });
 
     it('should display highlighted column selectable cells DataGrid with cell selection', () => {
-      $('[data-accessibility-id="26"]').click();
+      $('[data-accessibility-id="27"]').click();
       $('#root').moveTo({ xoffset: 0, yoffset: 0 });
       Terra.validates.element('highlighted column selectable cells cell selection', { selector: '#highlight-column-selectable-data-grid' });
     });